- add meson build helper (#8147)
