+++ conan/tools/meson/__init__.py
@@ -1,2 +1,3 @@
 # noinspection PyUnresolvedReferences
-from .meson import MesonToolchain
+from conan.tools.meson.toolchain import MesonToolchain
+from conan.tools.meson.meson import Meson
